+++ routes/html/script.js
